requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
rank-bm25
python-dotenv
pydantic
httpx
pytest
numpy
aiofiles
